Extracting logic for ensuring collection on an ExpandoObject
