+++ Source/Kernel/Events.Projections/ChildrenPropertyIsNotEnumerable.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Events.Projections.Changes
+namespace Cratis.Events.Projections
 {
     /// <summary>
     /// Exception that is thrown when a children property is not enumerable.
